workbook/requirements.txt
boto3==1.14.8
celery[sqs]==4.4.6
Django==3.0.7
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.5